refactor applib files
