+++ src/prov/api/healthcheck.py
@@ -1,7 +1,7 @@
 import json
 import falcon
 from prov.utils.logs import app_logger
-from prov.utils.graph_store import GraphStore
+from prov.applib.graph_store import GraphStore
 # from amqpstorm.management import ApiConnectionError
 # from amqpstorm.management import ApiError
 from amqpstorm.management import ManagementApi
